+++ samgraph/common/cpu/cpu_device.h
@@ -14,7 +14,7 @@ class CPUDevice final : public Device {
                        size_t alignment = kAllocAlignment) override;
   void FreeDataSpace(Context ctx, void *ptr) override;
   void *AllocWorkspace(Context ctx, size_t nbytes,
-                       size_t scale = Constant::kAllocScale) override;
+                       double scale = Constant::kAllocScale) override;
   void FreeWorkspace(Context ctx, void *ptr, size_t nbytes = 0) override;
   void CopyDataFromTo(const void *from, size_t from_offset, void *to,
                       size_t to_offset, size_t nbytes, Context ctx_from,
